change watcher and rebuilds

+++ addon/scripting_nodes/src/features/node_tree/code_gen/file_management/clear_addon.py
@@ -1,5 +1,7 @@
 import shutil
+import os
 
 
 def clear_addon_files(addon_path):
-    shutil.rmtree(addon_path)
+    if os.path.exists(addon_path):
+        shutil.rmtree(addon_path)
